Add desc, ul and ol to pizza and hotcakes html
It was added a description about what the recipe is about, an unordered
list with the ingredients for the food, an ordered list with the steps
to make corresponding food, and a format to add a future image of the
result. All the latest applied to pizza and hotcakes html.